Asset Manager Rewrite
- Louron Core -

Refactored and redesigned the Editor Asset Manager system. Previously there was the issue of managing handles based purely off their relative file path to the project asset directory, meaning that everytime you may have moved an asset in the assets directory, the handle would invalidate, and all sources in a scene that may have referred to that handle are now invalid. It would be hard to find all references of a UINT value and update it with the new handle...

SO I have opted to make it similar to Unity's workflow where every asset has a .meta file generated when first imported. This meta data file holds the static AssetHandle for that Asset once it has been imported. This means, we can move our assets around freely without worrying about invalidating handles that our scene reference, as handles will remain the same due to the .meta file that it alongside the asset, defining to us what the actual handle is, regardless of its new location in the assets directory.

Refactored Asset Importers to fit new system, most importantly, the Model Import.

Separated Asset Manager API into its own file.

Added SubmitToMainThread functionality to the engine. This is so we can submit functions to the main thread so that there aren't any issues with multithreading, data races, etc.

Implemented FocusOnEntity feature to EditorCamera class.

Implemented simple sphere VAO into Renderer for debug sphere rendering.

Fixed yet another issue with colliders, now it was their transformations in relation to their component properties such as box extents, center, and radius.

Updated TransformComponent::SetTransform functionality to decompose the local transform and set the internal properties to the passed transform.

Added functionality to Entity::AddComponent<BoxCollider> so that it sets the center and size of the collider as a close fit to the mesh filter attached to the entity.

Fixed Scene::DestroyEntity when trying to lock the mutex as a child when the parent had already locked the mutex... Ooops.

- Louron Editor -

Added more default resources.

Changed file watch library to EFSW as this has more functionality that I wasn't getting from filewatch.

Created more default options when creating entities in the hierarchy panel.

Implemented texture icons for content browser for compatible Texture2D assets.

Added debug line drawing for box colliders on the selected entity.

Added functionality to focus on entity when double clicked on scene viewport or hierarchy panel.