LOD Mesh Grouping
- Louron Core -

Implemented new component LODMeshComponent. This allows you to create an LOD system within the editor similar to Unity's implementation, but with my own twist which is as per usual, pree shitty.

The component allows you to prefer a custom distance over the cameras far plane, so if you wanted the range to be normalised within a custom distance of 0-100, then you can do that, otherwise if this is disabled.

Serialisation and deserialisation implemented for scene loading and saving.

Prefab creation and instantiation maps the correct internal scene UUID handle and internal prefab UUID handles to the entities that are listed in the LOD Element Levels.

Fixed FindChild of HierarchyComponent to actually search if its a child, and not just return it if it's in the scene.

- Louron Editor -

Create the LOD Mesh Component on a parent entity, have the LODs as children entities with the same transformation properties, assign them to the appropriate LOD level in the LOD Mesh Component, and away we go!

Modified functionality with Hierarchy Panel so you can click and drag entities without clicking into the entity.

Added functionality to add, modify and remove LODMeshComponent properties in the property panel.

Added +/- buttons for mesh renderer component materials.