Custom Shaders, Custom Materials, Custom Uniform Blocks for Materials, Scripting Integration, MSAA, Anisotropic Filtering
- Louron Core -

Deprecated the old ShaderLibrary, and made shaders assets. When the editor loads, it will load custom inbuilt shaders that are static and unchanging. This is handled with a new workflow for getting inbuilt shaders so the renderpipeline can use these during rendering.

Exteneded uniform Set functions in Shader class to include a larger variety of types.

Users can now create custom shaders in their project. There is a new workflow which allows users to create custom variables in what I'm calling a MaterialUniforms block in the shader.

When the shader loads, it will detect the appropriate part of the shader, and will read this in separate from the frag and vert shaders. This block will be inserted into both the vert and frag shaders before compilation. Users can access these custom uniforms through the MaterialUniforms u_MaterialUniforms block. If you declare a 'float myFloat;' in the custom MaterialUniform block, you can access it's members through doing 'u_MaterialUniforms.myFloat'.

This is cool and all now that we have a workflow to pass data into the custom shader, BUT how can we actually do this?

Material's have been updated to hold a new system called 'MaterialUniformBlock'. This is a way of managing these custom uniforms. Each material will have a uniform block with a map of uniforms with the key being the name, and then a GLSLType and UniformValue.

A GLSLType is a new enum which outlines all types we may use in custom shaders.

A UniformValue is a variant of all the internal C++ types for these GLSLtypes.

A MaterialUniformBlock is initialised from the shader, which means that it will create this uniform management system based on what the shader read in when compiling the shader code.

Now we have access to all our custom uniforms through a MaterialUniformBlock. How do we now use this! Well with MaterialUniformBlock::SetUniform() passing a uniform name, GLSL type and UniformValue ofcourse! This will store the uniform details and its value for the uniform in the material. This is good for default uniform values in the editor.

Did I mention it's all serialised and deserialised? Well if i didn't i have now...

So now we have a material, and the material has a layout of all the custom uniforms that are read back from the shader the material uses.

Now what if we want to change these values during runtime? Well we can directly set the uniform values to the 'master' material, and it will unilaterally apply these values to all meshes that use the 'master' material. Hmm doesn't sound too good, maybe in the editor when we are configuring base values for the material, but not during runtime when we want meshes to apply the changes ONLY to their mesh. E.g., a dissolve material. If we set the dissolve factor to the master material, EVERYTHING with the dissolve material will dissolve... We don't want this. We want to control these uniforms PER mesh.

SO! Our Mesh Renderers materials are now paired with a unique MaterialUniformBlock. This block will override the master material when rendering that specific piece of geometry.

This means, during the rendering pipeline when it sorts the meshes based on their material, it will now be keyed to the Material and the MaterialUniformBlock. If the MaterialUniformBlock paired with a Material in the MeshRenderer is nullptr, we will use the 'master' material values. However, if our pair in the MeshRenderer has a valid MaterialUniformBlock, we will create a new render path for this mesh, separating it from the default uniform values of the 'master' material, and applying its own uniforms!

This is unlike Unity that just creates an entirely new material, but I don't want to do this as we may have dangling memory when creating instances of materials that aren't really tied into the asset manager, and their lifetime extends that of the object.

I don't want that as these MaterialUniformBlock's will apply purely just to one MeshRendererComponent, so when the MeshRendererComponent is removed, so too is this custom MaterialUniformBlock. Ezpz :)

Now that we have overridable uniform blocks for our mesh renderer materials, how can we then set these values? The Script Core API has been extended largely to include new types such as Material, MaterialUniformBlock, and MeshRendererComponent. The workflow in a script will be that shown in the example project scripts "Material Uniforms Test.cs". First get the MeshRendererComponent, then enable custom uniform blocks for the appropriate material index on the component, then GET this uniform block and store it in a variable that will extend the lifetime of the script (probs not best to get every frame).

Then once your have this uniform block, you can SET all the uniform information you need.

See just a small commit...

But I have used this workflow to create a custom dissolve shader which dissolves a mesh using a noise map.

In the script example, you can see the pingponging of the dissolve shader attached to the dissolve material in action!

Now onto other improvements

Redesigned the asset manager alittle to also include InBuiltAssets, which are static assets the editor provides, and are immutable to the user such as default inbuilt shaders, inbuilt FBX models, inbuilt textures, etc. When I implement binary serialisation for distributing a game, all these inbuilt assets will be shipping inside the binary file alongside all assets created and added by the user.

I updated the FrameBuffer class to include support for multisampling. This did not work with the entity picking texture, so I opted to switch that to an SSBO, and do manual depth testing in the shader to determine what entity is at the forefront, and setting the entity uuid value in the SSBO. When reading back, we just read from the SSBO, simples.

Also updated the GLAD loader to include all extensions, including Anisotropic filtering.

Updated the Texture loading so we actually load the textures properly now. There were artifacts from not using the correct filtering on mipmap levels, so normals were super glitchy and looked weird. Need to consider loading HDR textures aswell in the future maybe? Unsure about how to make this work with the shaders.

Added step to render pipeline. If shader is not compiled successfully, it will display an invalid texture on the object.

Massively cleaned up the master shader so it is versatile with making custom shader templates.

- Louron Editor -

Implemented Material Properties Window. This is so you can easily update information pertaining to a custom serialised material. Most of the functionality is there to set default uniform values for the shader, drag drop shader onto shader field, change name of material, change textures, etc.

Added functionality to create new shader which copies the default shader to the target location.

Updated Properties Panel to incorporate changes to shaders and materials. Also added functionality to auto use a colour selector when the script variable name includes Colour, or Color, RGB, etc.

- Louron Script Core -

Implemented material API to access Materials, MaterialUniformBlocks, and Mesh Renderers.

See example script in this commit on how to use this system.