Occlusion Culling (Naive Implementation)
- Louron Core -

Implemented occlusion culling into the forward plus pipeline.

I ideally would want to reduce the number of depth calls if it is not visible, however, this is not possible as the depth drives the occlusion testing.

If I were to not include an object in the depth pass because it is not visible, I will never know again if it is visible or not.

Had to remove multithreading for the depth pass sorting to achieve this.

Pretty happy with it honestly, thinking of maybe making it a toggleable option that can be used in dense areas to increase rendering performance.

The only struggle is with rendering the depth buffer, which can be quite expensive... Need to just git gud i'd say.

Expanded the render stats.

- Louron Editor -

Visualise new render stats in Rendering Stats tree node. You can see here how many objects are being occluded, thus not included for the colour pass.

Fixed issue with material inspector in the properties panel where it'd bug out when multiple materials of the same name are added to the treenode layout. Needs a unique label, so just added the loop iterator to the end after ##.