CSM Changes, Depth Flag for Materials, Cleanup
- Louron Core -

Fixed issue where you couldn't mouse pick transparent objects.

Added WriteDepth toggle to material. This is only available for transparent objects. This means that when the object is rendererd, it will either write the depth values to the buffer, or not. This is good for some transparent objects that might be semi transparent, and you don't want to see pixels behind opaque parts of the geometry that are not transparent. Please make sure that you DISCARD transparent pixels when they are not visible in order to see other transparent objects behind it if the transparent object is set to write the depth. If you only set the alpha channel and not discard, it will block transparent objects behind that transparent zone from being rendered.

Altered CSM to dynamically choose the plane distances.

Added feature to directional light so that you can determine the max viewable distance of the shadow. This yields performance and ensures quality shadows up close.

Fixed other miscellaneous bugs... There are so many of them!

- Louron Editor -

Added functionality to set max viewable shadow distance in DL component in property panel.

Added functionality to toggle depth write for transparent materials in material property editor.