Material Cleanup
- Louron Core -

Got rid of the old Material base class and just replaced it with the PBRMaterial derived.

Updated all the functionality for Material and SkyboxMaterial.

Skyboxes only serialise through editor now, and do not serialise via the engine, similar to how Material's do it.

Added equality operator to Asset base class to check handles.

Tried better caching optimisation in render pass and render sorting, only got .1ms increase, so I won't try again for awhile...

When making copies of VertexArray, we can set if that VertexArray will delete the underlying VAO on destruction or not. VertexArrays could be copied, so we need to make sure that when a copy is initialised, we set it to not delete on destructor.

- Louron Editor -

Added functionality to serialise skybox when modified in editor.

Changed WriteDepth bool in Material to just be another flag that can be set for the render_type.