Transparency on Material Import & Point Light Shadow Fix
- Louron Core -

Added functionality to read transparency of material during material import for mesh. Updates the render type of the material and sets the alpha channel of the colour to 1.0f - transparency factor. This ensures the alpha is set, and it is processed in the transparency pass.

Fixed issue where the full sphere of a point light was not including all entities from the Octree query.