Major Rendering Re-Write - Multiple Camera Support
- Louron Core -

Re-wrote the entire processing logic of the Rendering Pipeline. Rendering paths are now static pipelines, opposed to being created per Scene.

F+ pipeline now has support to render multiple camera targets and display them to the main viewport.

F+ Pipeline has a static unordered map of Scene Contexts with the Scene raw pointer being the key, and the context being the value. Each Scene Context has a similar setup for Camera Contexts, and where there are multiple cameras in the scene, there may be multiple camera contexts within the scene context.

This is used to create a separation of the data required between the global scene, and the local camera context. E.g., point and spot lights, and bone transformations are scene global, whereas the indices of point lights due to tile based culling is camera specific, or directional cascaded shadow maps are dependent on the cameras frustum, so is kept within the camera context.

Each scene pass we gather cameras, we then sort the render queue using Octree frustum culling and Query driven occlusion culling, to fill out the Depth, Opaque, and Transparent queues. Then shadow maps are updated, SSBOs are updated.

Each camera pass the depth pass is run, tiled based light culling is computed, occlusion culling is performed, and main colour rendeirng is then performed before looping through the next camera.

Camera Components have the option now to set the depth of the camera, the viewport (normalised values), and if it is to be composed in the final screen image (e.g., maybe you want to render a camera, but do not want it to show in the final viewport, rather you can use the rendered image as a texture for something else? Maybe a CCTV camera is rendered, then you use the final image as a texture input for a HUD on an in game element, such as a computer screen, need to implement the logic behind this so scripts can allow this to happen).

Decreased max UUID to UINT32 MAX -2, this allows to have another unique identifier in the behind the scenes so that the ECS does not assign that value, and I can use that identifier as a key. E.g., in the rendering pipeline now I assign this value to the Editor Camera camera context KEY. This is so for whatever reason RNG may have to create an identical value, it won't be the same as the new identifier UINT32 MAX - 2, and at the same time, it is not a NULL_UUID, so its sort of a VALID identifier for engine purposes, but invalid for ECS purposes.

EditorCamera class now has its own FrameBuffer.

CameraComponents now have new options, and its own FrameBuffer.

Moved the Skybox Vvertex Array out of the component, this was creating a new skybox VAO for each SkyboxComponent you may have in the ECS... Stupid I know! This now resides in the static Renderer along with the DebugCube VAO.

Added CheckOpenGLError to Renderer API.

Moved Sphere v Frustum intersection checking into Frustum class.

Scene rendering logic is now separated from the OnUpdate function. Scene will call the Render Pipeline, then it will composite the final images into the scene's framebuffer using the viewports of the CameraComponents.

- Louron Editor -

Added functionality to properties panel to change the new Camera Component data.

Added Camera Preview to main viewport in editor when selecting a camera in the hierarchy.

Modified editor code to meet new changes to core engine.

Included functionality to hide "hidden" files and folders on the OS from the content browser panel.

- Louron Script Core -

Ran code cleanup to clean up C# code.