Thread Pooling Job System
- Louron Core -

Implemented rudimentary thread pool system for tasks that can utilise the full extent of the CPU's potential.

Jobs are kicked to the jobsystem which sit in a global queue protected by a mutex and has a condition variable to ensure threads aren't busy waiting and spinning waiting for a job when there are no jobs for that thread.

Workers then pop a job from the thread queue and execute their jobs with profiling.

Jobs can also have a counter which is basically a mechanism that allows us to know when the job has completed. This counter has the capability to wait for jobs to be completed.

Jobs can be submitted either as non-persistent or persistent. Non-persistent jobs are to be completed in the current frame to avoid stacking up jobs over multiple frames where they should only be limited to a certain frame. Persistent jobs exist past the end of the frame - it is recommended to use a counter to synchronise persistent jobs, e.g., kick a job from on frame, and wait on its completion before kicking the same job for the next frame.

Each worker has a local queue that other threads can steal jobs from, however, submission to local queue has not yet been implemented as all jobs are submitted to the global queue, maybe workers should only submit jobs to their own local queue, then other workers can steal from it?

TODO: job parking/yielding, e.g., if we need to wait for all child jobs to complete, we do not want to have the entire thread sleep which all child jobs are completed, we should yield the parent job, and then walk through local queue?

Updated logging functionality to maintain a vector of all logging strings, this is so we can print console output to an ImGui window. Also included a new L_SCRIPT_type macro for scripting logs.

Implemented entity duplication logic. TODO: need to fix transformation hierarchy copying.

Implemented first use case of job system for processing animations. On my system, I can process 300 animated characters in the frustum with a frame rate of 60FPS.

- Louron Editor -

Implemented GUI to visualise thread pool profiling.

Implemented GUI to visualise console outputs with filtering.