Scripting Rework - Native C++ Scripting
Removed C# scripting and replace with C++ scripting solution.

Doxygen HTML can be found in Louron Native Script Core -> doxy-gen -> html -> index.html

Basically the C++ scripting solution is as follows:
 - User creates Scripting Project, Script Core API includes are moved into Scripts/Script Core API folder.
 - User creates Scripts through editor which pumps out template script.
 - Editor detects changes and uses clang-c to parse the header files and identify all the exposed fields in the script class.
 - Editor will then keep this information, and generate a CPP file to be included in compilation. This CPP registers all the scripts meta data into a global unordered_map which is queried by the host application (engine) when it is loaded.
 - Engine will invoke std::system call to compile Script Code DLL.
 - Engine will reload assembly and cleanup as necessary each hot reload.
 - Engine will get meta data, factory function ptrs, release function ptrs, field names, field offsets, field types, etc., and put this into its internal data structures.
 - Engine can then call the appropriate IScript functions directly invoking either the base class, or the overriden class in the DLL.

All script instances are kept on the DLL side as unique ptrs. When we use the factory function for a script on an entity, it will create a new unique_ptr of the IScript on the DLL side, and pass a raw ptr back to the engine.

When the script instance is destroyed by the EXE, it calls a release function ptr in the DLL to release the memory on that side of the ABI boundary.

Script Register registers all functions within the ENGINE side which will be called by the DLL.

Macros in the Script Core API make it possible to communicate with the ENGINE safetly.

Still needs a cleanup of the system to remove duplicated functionality/redundancy.

Script Register needs a severely close look at.

Script Core API source of truth is in the Louron Native Script Core project. The copies found in Editor/Resources/Script Core API, and Sandbox Project/Scripts/Script Core API are merely copies for distribution and compilation requirements. E.g., when we compile the main engine and editor, it copies all the header and CPP files from the Louron Native Script Core project into Editor/Resources. This ensures most up to date copy in the editor resources.

When the editor is distributed, it will remain in the resources.

This is so that it can then again copy into Project/Scripts/Script Core API for now COMPILATION and EDITING purposes. I want the user experience to be at their finger tips so they are swimming in a sea of sqwiggly red lines as it doesn't have any proper includes, so everything is available then in the Project/Scripts/Script Core API/ for them to use. Obviously during compiling this include header is overriden with the source of truth to ensure absolute correctness and no silly business.